Fixed code redundancy issue in SQLAdapter.
